Add playRound function
Determine if the player or the computer wins. Convert user entry to
lower case to sanitize. Add a helper function that capitalizes the first
letter of a string for ease of output.

This version contains commented out code for the playRound function
showing alternate ways it can be written. I first used nested if/else
statements but later shortened using ternary operators.